move httpclient test data to separate package
